bacontree functions implemented
BaconTree class's functions have been implemented. A quick test has been done to verify the code works.

class TreeNode* getTreeNode(string actorName);
A new method has been added to BaconTree. It takes a string containing an actor's name as a parameter and returns a pointer to that actor in the tree, or NULL if the actor is not in the tree. This will make it easier to look up actors within the tree, as otherwise it requires a number of pointer dereferences and function calls.

TreeNode.parent has been made private to keep clients from modifying it. Only the BaconTree should need to touch it, and it has been listed as a 'friend' to facilitate this.